use f64 instead of f32 (juniper only supports f64)

+++ traffic/src/lib.rs
@@ -5,11 +5,11 @@ use juniper;
 #[derive(Serialize, Deserialize, Copy, Clone, Debug)]
 pub struct TrafficYearData {
     pub year: i32,
-    pub dhv: f32,
-    pub directional_split: f32,
+    pub dhv: f64,
+    pub directional_split: f64,
     pub aadt: i32,
-    pub aadt_yearly_change: f32,
-    pub aadt_10_year_change: Option<f32>,
+    pub aadt_yearly_change: f64,
+    pub aadt_10_year_change: Option<f64>,
     pub sadt: i32,
     pub sawdt: i32,
     pub wadt: i32,
@@ -33,8 +33,8 @@ pub struct TrafficData {
     pub hwy_type: String,
     pub location_desc: String,
     pub reg: String,
-    pub section_length: f32,
-    pub connecting_link_length: f32,
+    pub section_length: f64,
+    pub connecting_link_length: f64,
     pub secondary_desc: String,
     pub travel_patterns: HashMap<String, Vec<TrafficYearData>>,
 }
